Se agrego el submit en la pantalla de myprofile con funciones de javascript. 	modified:   app/controllers/display_controller.rb 	modified:   app/models/usuario.rb 	modified:   app/views/display/myprofile.html.erb 	modified:   app/views/zonas/show.html.erb 	modified:   db/migrate/20111208000111_create_zonas.rb

+++ db/migrate/20111208000111_create_zonas.rb
@@ -5,7 +5,7 @@ class CreateZonas < ActiveRecord::Migration
       t.string :nombre
       t.float :longitude
       t.float :latitude
-      t.integer :radio
+      t.float :radio
 
       t.timestamps
     end
